# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-root",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `root(value, n)` by composing two primitives.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. ask `srvcs-floatdivide` for `exp = 1 / n`;\n2. ask `srvcs-floatpower` for `result = value ^ exp` — i.e.\n   `root(value, n) = value ^ (1 / n)`.\n\n`srvcs-floatpower` returns `422` if the result is not real (e.g. an even\nroot of a negative value); that `422` is forwarded. If a dependency is\nunreachable it reports itself degraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value",
          "n"
        ],
        "properties": {
          "n": {
            "type": "number",
            "format": "double"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "RootResponse": {
        "type": "object",
        "required": [
          "value",
          "n",
          "result"
        ],
        "properties": {
          "n": {
            "type": "number",
            "format": "double"
          },
          "result": {
            "type": "number",
            "format": "double"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}